Update docker-compose.yml
SonarCloud is correctly flagging the default password in docker-compose.yml. We added a comment in the code to justify the false positive.

This is a docker-compose.yml file, which is not used in production. It must have a default value for local setup, but that default is flagged as a risk.